requirements-local.txt
ansi2html==1.8.0
boto3==1.28.49
botocore==1.31.49
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
dash==2.13.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
duckdb==0.8.1
Flask==2.2.5
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
MarkupSafe==2.1.3
nest-asyncio==1.5.8
numpy==1.25.2
packaging==23.1
pandas==2.0.3
plotly==5.17.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
retrying==1.3.4
s3transfer==0.6.2
six==1.16.0
tenacity==8.2.3
typing_extensions==4.8.0
tzdata==2023.3
urllib3==1.26.16
Werkzeug==2.2.3
